Revert "Refs [Task][#445482][component view my course with student role] [Task][#445479][route get my courses with student role]"
