+++ client/src/_services/student.service.js
@@ -11,8 +11,7 @@ export const studentService = {
   getSubjectsOfCourse,
   getClassesOfSubject,
   getSubject,
-  getMyClasses,
-  getMyCourses
+  getMyClasses
 };
 
 async function getAllCourses() {
@@ -89,7 +88,7 @@ async function getAllSubjects() {
 
 async function getMyClasses() {
   try {
-    let respone = await axios.get(`${process.env.VUE_APP_API_BACKEND}/student/myClasses`, {
+    let respone = await axios.get(`${process.env.VUE_APP_API_BACKEND}/account/me/myClasses`, {
       headers: authHeader()
     });
     return respone.data.classes;
@@ -97,18 +96,6 @@ async function getMyClasses() {
     throw error;
   }
 }
-
-async function getMyCourses() {
-  try {
-    let respone = await axios.get(`${process.env.VUE_APP_API_BACKEND}/student/myCourses`, {
-      headers: authHeader()
-    });
-    return respone.data.courses;
-  } catch (error) {
-    throw error;
-  }
-}
-
 async function registerCourse(courseId) {
   try {
     let response = await axios.post(
